refactor(user): Move GymmateUser Files
