--- src/main/java/org/helloworld/gymmate/domain/user/member/service/GymmateUserDetailsService.java
@@ -1,4 +1,0 @@
-package org.helloworld.gymmate.domain.user.member.service;
-
-public class GymmateUserDetailsService {
-}
