+++ src/main/java/org/helloworld/gymmate/domain/user/service/GymmateUserDetailsService.java
@@ -1,0 +1,4 @@
+package org.helloworld.gymmate.domain.user.service;
+
+public class GymmateUserDetailsService {
+}
