+++ src/main/java/org/helloworld/gymmate/security/authentication/GymmateUserDetails.java
@@ -2,7 +2,7 @@ package org.helloworld.gymmate.security.authentication;
 
 import java.util.Collection;
 
-import org.helloworld.gymmate.domain.user.member.model.GymmateUser;
+import org.helloworld.gymmate.domain.user.model.GymmateUser;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.userdetails.UserDetails;
