+++ src/main/java/org/helloworld/gymmate/security/oauth2/LoginSuccessHandler.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import org.helloworld.gymmate.common.cookie.CookieManager;
 import org.helloworld.gymmate.common.enums.TokenType;
-import org.helloworld.gymmate.domain.user.member.service.GymmateUserService;
+import org.helloworld.gymmate.domain.user.service.GymmateUserService;
 import org.helloworld.gymmate.security.policy.ExpirationPolicy;
 import org.helloworld.gymmate.security.token.JwtManager;
 import org.springframework.beans.factory.annotation.Value;
